local search bug fixed, height computation inside greedy procedure and local search before optimize

+++ include/problem_utils.hpp
@@ -36,7 +36,7 @@ public:
     int source_radius;                  // the source radius if there is one, or the graph`s radius
     map<DNode, map<DNode, Arc>> arc_map;// used for fast arc lookup
     DNodeIntMap node_height;
-    double solution_height;
+    int solution_height;
 };
 
 void PrintInstanceInfo(Problem_Instance &P);
